Improvements to Test Cases
This patch begins to get test cases closer to passing. The biggest change, overall, is to get the ethereum common tests directly in the root of the umbrella app and make sure each sub-app appropriately references these tests (it used to be pulled in by each sub-app individually, which makes less sense now that we have an umbrella app). Sadly, a few of the different apps relied on having sightly different versions of the Ethereum Common Tests and thus need to be updated. Finally, the tests have changed in many cases, so we have some fixing up to do.

This patch, anyway, gets the tests running again in each app. We have some work to do getting the test cases fully passing by each app thereafter.