Add CircleCI Badge
This patch adds a badge for CircleCI. As a separate PR, we should consider breaking the test cases up into separate tasks, but currently they should run pretty quickly, so this may not be necessary. The one advantage it might give us is to show exactly which sub-projects are passing or failing.