Add ABI Specification Library
This patch adds an ABI specification parser. This allows users to parse the ABI that comes out of Solidity and other Ethereum compilers and use it with the ABI library.

Overall, the ABI library needs a large re-factor, since many of the interfaces are inconsistent and there are plenty of usability issues. Overall, I think `FunctionSelector` should go away (or be reduced in scope) and we should consider using `ABI.Spec` as our proper specification. This will come in due time, but this is a first-start to making this library useful for a larger population.